refactor: update structure and logger
