+++ .env
@@ -13,7 +13,7 @@ TEST_MODE=auth
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
 POSTGRES_DB=music_db
-POSTGRES_PORT=5432
+POSTGRES_PORT=6432
 #POSTGRES_HOST=postgres
 POSTGRES_HOST=localhost
