+++ docker-compose.db.yml
@@ -8,7 +8,7 @@ services:
       POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
       POSTGRES_DB: ${POSTGRES_DB}
     ports:
-      - '5432:5432'
+      - '6432:5432'
     volumes:
       - postgres-data:/var/lib/postgresql/data
