+++ src/database/repositories/track.repository.ts
@@ -4,7 +4,7 @@ import {
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
-import { ErrorTrackMessages } from '../../shared/constants/error-messages.constants';
+import { ErrorTrackMessages } from '../../common/constants/error-messages.constants';
 import { CreateTrackDto } from '../../modules/track/dtos/create-track.dto';
 import { UpdateTrackDto } from '../../modules/track/dtos/update-track.dto';
 import { Track } from '@prisma/client';
