+++ src/database/repositories/user.repository.ts
@@ -8,7 +8,7 @@ import {
 } from '@nestjs/common';
 import { CreateUserDto } from '../../modules/user/dtos/create-user.dto';
 import { UpdatePasswordDto } from '../../modules/user/dtos/update-password.dto';
-import { ErrorUserMessages } from '../../shared/constants/error-messages.constants';
+import { ErrorUserMessages } from '../../common/constants/error-messages.constants';
 import { User } from '@prisma/client';
 import { DatabaseService } from '../database.service';
