+++ src/modules/album/album.module.ts
@@ -5,7 +5,7 @@ import { AlbumRepository } from '../../database/repositories/album.repository';
 import { TrackModule } from '../track/track.module';
 import { FavoritesModule } from '../favorites/favorites.module';
 import { DatabaseModule } from '../../database/database.module';
-import { ErrorModule } from '../../shared/error/error.module';
+import { ErrorModule } from '../../common/error/error.module';
 
 @Module({
   imports: [
