+++ src/modules/artist/artist.module.ts
@@ -6,7 +6,7 @@ import { AlbumModule } from '../album/album.module';
 import { ArtistRepository } from '../../database/repositories/artist.repository';
 import { FavoritesModule } from '../favorites/favorites.module';
 import { DatabaseModule } from '../../database/database.module';
-import { ErrorModule } from '../../shared/error/error.module';
+import { ErrorModule } from '../../common/error/error.module';
 
 @Module({
   imports: [
