+++ src/modules/favorites/favorites.module.ts
@@ -6,7 +6,7 @@ import { TrackModule } from '../track/track.module';
 import { ArtistModule } from '../artist/artist.module';
 import { AlbumModule } from '../album/album.module';
 import { DatabaseModule } from '../../database/database.module';
-import { ErrorModule } from '../../shared/error/error.module';
+import { ErrorModule } from '../../common/error/error.module';
 
 @Module({
   imports: [
