+++ src/modules/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserController } from './user.controller';
 import { UserRepository } from '../../database/repositories/user.repository';
 import { UserService } from './user.service';
 import { DatabaseModule } from '../../database/database.module';
-import { ErrorModule } from '../../shared/error/error.module';
+import { ErrorModule } from '../../common/error/error.module';
 
 @Module({
   imports: [DatabaseModule, ErrorModule],
